Adds gitea README and installation user-data in CF template
I added a little note on accessing the Gitea server after deployment
and added a Gitea installation BASH script to the EC2 user-data in the
Gitea CF template. The script is still a WIP, since I'd like to
preconfigure Gitea with a configuration file in the script rather than
via the web interface.