Adds simple git server with git-annex for large files
As an alternative to running something like Gitea, it's simpler and
less resource intensive to use `git --bare`, which is what I've done
here. It's a basic EC2 instance with a `git` user and a dynamic number
of bare git repositories created under the `/home/git` directory. Each
repository is also initialized with git-annex for large file storage.
SSH public keys are added to the `git` user's authorized_keys file for
access.